Move `submit_batches_from_source` to CLI
Remove the `submit_batches_from_source` function from the "runner"
module and add it to the "playlist" module in the CLI.
